refactor: move field enums to dto package
