+++ src/main/kotlin/se/uulm/snowballr/backend/model/dto/criterion/CriterionField.kt
@@ -1,4 +1,4 @@
-package se.uulm.snowballr.backend.model.incoming.criterion
+package se.uulm.snowballr.backend.model.dto.criterion
 
 enum class CriterionField(val grpcPath: String) {
     TAG("criterion.tag"),
